commit on RamanBranch

+++ ERP/src/main/java/com/ERP/repositories/SalaryStructureRepository.java
@@ -1,5 +1,6 @@
 package com.ERP.repositories;
 
+import com.ERP.dtos.SalaryStructureDto;
 import com.ERP.entities.SalaryStructure;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SalaryStructureRepository extends JpaRepository<SalaryStructure, Integer> {
+public interface SalaryStructureRepository extends JpaRepository<SalaryStructure, Long> {
     public List<SalaryStructure> findAllByRole(String role);
 }
